Decouple LoadBalancerControlService from DeploymentService
