+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 
 file.upload-dir=ServiceJars
-deploy.HAProxyHostname=localhost
-deploy.HAProxyPort=9999
+load-balancer-control.HAProxyHostname=localhost
+load-balancer-control.HAProxyPort=9999
